requirements.txt
annotated-types==0.6.0
attrs==23.2.0
blinker==1.7.0
click==8.1.7
distlib==0.3.8
filelock==3.13.1
flasgger==0.9.5
Flask==3.0.2
Flask-Cors==4.0.0
flask-openapi3==3.1.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mistune==3.0.2
platformdirs==4.2.0
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
setuptools==69.5.0
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
virtualenv==20.25.0
Werkzeug==3.0.1
